--- src/main/java/subway/favorite/ui/FavoriteExceptionControllerAdvice.java
@@ -1,16 +1,0 @@
-package subway.favorite.ui;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import subway.favorite.exception.FavoriteNotFoundException;
-
-@RestControllerAdvice
-public class FavoriteExceptionControllerAdvice {
-
-    @ExceptionHandler(FavoriteNotFoundException.class)
-    public ResponseEntity<Void> NullFavorite() {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-    }
-}
